Adding using.inc to handle godot namespace.
GDExtension has everything under godot namespace. In engine no such namespace exist.
Since it is an error in C++ to use namespace which is not defined preprocessing is
used to detect should godot namespace be used.

To avoid having to write this check to every header "using.inc" is added. This way
using.inc only needs to be added every header without worrying about details.

register_types is specially handled as gdextension_interface.h gives information in
the GDExtension which is given modules/register_module_types.h in engine.