Header remapper.
These headers mostly redirect the include the engines location for asked class. GDExtension is prioritize for stability and ease of use.

Two of the headers do something special. math.hpp needs to define some functions and macros them to exist. UtilityFunctions does not exist
in-engine. VariantUtility has the method with little bit different parameters on some but they can used.